check: Always write parsing results to 'stdout'
Issues: #4

The 'check' command focuses on parsing the JASS scripts. As such, a
parsing failure is not exactly an error, per sey. It does not make sense
to display the results to 'stderr'.

This behavior has been changed, and now the results display to 'stdout'.
The command will still exit with `EXIT_FAILURE` when parsing fails,
however. Note that other errors will continue to display to 'stderr'.