common: Add `command` argument to `initialize`
This forces the command to be passed directly to the function, rather
than attempting to source it from `arg [0]`. This addresses a concern on
Windows systems when forward slashes are used to execute a command (e.g.
`lua map/imports`). The function `Path.base_name ()` only acts upon the
operating system dependent directory separator (as specified internally
by Lua), and thus would not behave as intended.