map: Remove unused files/tools
This is a pretty large removal.  For starters, we no longer include
third party tools within **Map**.  This means that Grimex, MPQEditor,
and PJass are all being completely removed.  On top of that, all project
documentation has been inlined into the README.  And finally, a number
of files/commands are just redundant given the large refactoring.