check: Switch from PJass to Wurst
This is the beginning of dropping PJass support entirely from Map.  Yes,
it makes checking the scripts slower.  However, there isn't another
choice given that we want to deal with depdency ordering at the function
level automatically (and, Map does not plan to implement such behavior
in-house at this time).