task/build: Stop inlining strings in Jass scripts
This is more of an issue with the current design than a desire to stop
actually inlining the strings.  Previously, we were processing the
scripts within Map.  This allowed us to easily inline trigger strings.

However, now we are passing a list of files to Wurst for handling.
Wurst will have no knowledge of the trigger strings, and as such cannot
make a decision to inline or not.