map: Rework source file handling
Issues: #40

In light of Wurst's recent handling of the project folder, the `scripts`
setting has been removed and replaced by `source`.  This will force the
user to explicitly define the `source.directory` used to contain the
Jass/Wurst files used by a project.  And in the case of external Jass
files, a means to deal with this (`source.include`) has been added.