[Map] Automatically handle Wurst related files
Automaticaly handling of Wurst related paths is now extended to the
following:

- `wurst`
- `wurst.dependencies`

In the first case, if `source.directory` is not equal to 'wurst', then a
symbolic link be created.  This is necessary as various functionality of
Wurst depends on that hardcoded directory path.  In the second, any
entries within `source.wurst.dependencies` will be used to populate the
file.