[Luacheck] Add a WC3 standard
This custom standard for Luacheck represents the Lua environment
presented by Warcraft III.  All Jass identifiers are present (including
some that are not within the usual files).  The Lua 5.3 standard has
been updated to reflect the limitations imposed on the user.  And, the
`require` global has been 'added' by Map.

Additionally, a `luacheckrc` file is provided that is used as the
`--default-config` option for Luacheck.