IMP: Enhance (unknown) byte handling
Issue: #64

This commit returns the value of the byte that exists before file paths
on unpacking, and allows the user to specify a value upon packing.
Whereas previously this value was determined via the path itself (i.e.
whether or not it included `war3mapImported\\`), it now defaults to the
value used for all imports as of patch 1.32 (i.e. `0x1D`).