W3X: Enhance import handling
Issue: #64

An internal list of files ignored as imports by the World Editor is now
used to identify when the `war3map.imp` needs to be updated.  This list
may not be exhaustive, and supports patch 1.32.  This support allows
removal of the `import` argument in functions that used it.

Additionally, the abilility to pass options when creating a W3X object
has been added.  Currently, this only amounts to providing the import
byte for the `war3map.imp`.