Map: Introduce W3X directory support
Issue: #66
Issue: #70

This commit introduces W3X directory support.  For the user, this
primarily means that the input map can be either a W3X directory or MPQ.
Existing handling remains consistent.  As for the output map, a MPQ
archive will remain the default.  But by specifying an option flag it is
possible to output to a W3X directory.