Script: Change default debug mode to 'name'
It makes more sense to have 'name' be the default.  The biggest reason
is that it exposes less of the user's filesystem hierarchy to the built
map.  Furthermore, the user is expected to understand the behavior of
the `package.path`.  As such, having the name of the required module (as
opposed to file path) should be enough.