Settings: Introduce `package.cpath` for build files
Issue: #83

This now gives the full control over the package paths for user build
files.  This lack of functionality for the C path was an oversight.

Additionally, there is now a distinct split between the required modules
used within the build files and those used within Map itself.  This is
the extent of the sandboxing used.  Perhaps at a later time a more
robust method can be utilized.