Imports: Refactor design
Issue: #85

This reworks imports, both internally and in how the user interacts with
them.  Previously, the only user interaction was defining imports in the
build environment.  They had no knowledge of what was in the input map.

Now, the imports table is populated with all files in the map, except
for those exposed to the build environment itself.  This can give the
user the ability to exclude certain imports from the output archive.