Settings: Make the `script` table optional
Issue: #74

Previously, the `script` table was required.  The expected use of Map
was to generate a `war3map.lua`.  However, with additional support for
various map files added, it is possible that a user may want a
configuration file to only run build files and package the output
archive.  And now that is possible.