Modules: Detect `require` at start of file
Trying to be too smart was the issue.  Yes, we had a nifty way to detect
'valid' uses of `require`.  This method leveraged LPeg's `B`, which
checked fixed strings immediately before the current position.  However,
when `require` is at the start of a file there are no fixed strings
immediately before, so the match would fail.

The fix is to check for a single usage of `require` immediately after
the shebang, which if absent will in turn represent the beginning of the
file.  For subsequently uses, we ensure that `require` is prepended by
valid characters.