Modules: Adjust behavior and cleanup
The output format has changed.  Now a list sorted by module names is
returned, with each element containing the module name, path, and, by
default, the module contents.  These modules are iterable via `ipairs`
(i.e. `[1, N]`).  This is a breaking change.

Additionally, the root module is now specified by name instead of path,
and an attempt is made to find it within the provided path.  In the
output, `modules [0]` will represent the root module.