Lua: Add `war3map.lua` support
Allow the ability to `pack` the `war3map.lua`, using the following
format to specify modules:

```lua
local input = {
    {
        name = '',
        path = '',
        contents = ''
    },
    -- More modules.
}
```

Modules are expected to be iterable via `ipairs` (i.e. `[1, N]`).  Note
that `modules [0]` will represent the root object, and will be included
at the end of the output file.  The root object should probably contain
`main` and `config`.

Note that the output of `Modules.load` can be passed directly to this
module to produce a usable `war3map.lua`.