INI: Switch to using LPEG for `unpack`
Issue: #99

Switching to LPEG should make it easier to properly support the file
format.  In particular, double quoted lists are now handled properly.
Additionally, fewer assumptions regarding values are made.

For unquoted values, everything until the EOL is now returned.  Within
the game files there are cases where trailing comments are not removed.
There are cases where they are.  How are we supposed to know?

For quoted values, both single and double, the quotes are no longer
removed.  There are cases where some values must be double quoted.
Others must be singled quoted.  Again, how are we supposed to know?

So we make less assumptions, and let the user decide.