Units DOO: Update and improve
Issue: #99

Now attempt to automatically determine if the file format is for
Reforged (1.32).  This is done by looking at the flags value.  If it is
a printable character (i.e. greater than `0x20`), then we make the
assumption that it is the latest file format.  Previously, we required
the user to supply a version table.

Furthermore, we adjust a few things, including the output format and
introducing `position` for `unpack`.  This could be considered a
breaking change.