Doodads DOO: Update and improve
Issue: #99

As with the unit version, we now automatically detect the file format
introduced with Reforged (1.32).  This is done by looking at the byte
that either represents a doodad's flags or skin.  If this byte is a
printable character (i.e. greater than `0x20`), we assume it is a skin,
and thus the latest format.

Other changes include support of `position` in `unpack` and adjustments
to the output format.